+++ src/main/java/org/realityforge/guiceyloops/server/glassfish/AbstractProvisioner.java
@@ -1,5 +1,6 @@
 package org.realityforge.guiceyloops.server.glassfish;
 
+import java.util.Map;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 
@@ -7,11 +8,16 @@ public abstract class AbstractProvisioner
   implements Provisioner
 {
   private String _baseHttpURL;
+  private Map<String, Object> _services;
+  private GlassFishContainer _glassfish;
 
-  public final void provision( @Nonnull final GlassFishContainer glassfish )
+  @Override
+  public void provision( @Nonnull final GlassFishContainer glassfish, @Nonnull final Map<String, Object> services )
     throws Exception
   {
-    _baseHttpURL = glassfish.getBaseHttpURL();
+    _glassfish = glassfish;
+    _services = services;
+    _baseHttpURL = _glassfish.getBaseHttpURL();
     configure( glassfish );
     deploy( glassfish );
     postDeploy( glassfish );
@@ -32,5 +38,13 @@ public abstract class AbstractProvisioner
   protected final String getBaseHttpURL()
   {
     return Objects.requireNonNull( _baseHttpURL );
+  }
+
+  @Nonnull
+  public final OpenMQContainer getOpenMQContainer()
+  {
+    final OpenMQContainer _openMQContainer = (OpenMQContainer) _services.get( OpenMQContainer.class.getName() );
+    assert null != _openMQContainer;
+    return _openMQContainer;
   }
 }