+++ src/main/java/org/realityforge/guiceyloops/server/glassfish/Provisioner.java
@@ -1,9 +1,10 @@
 package org.realityforge.guiceyloops.server.glassfish;
 
+import java.util.Map;
 import javax.annotation.Nonnull;
 
 public interface Provisioner
 {
-  void provision( @Nonnull GlassFishContainer glassfish )
+  void provision( @Nonnull GlassFishContainer glassfish, @Nonnull Map<String, Object> services )
     throws Exception;
 }